Create input component group, rename frontendcomponents page to Demo
