+++ web/components/app/App.js
@@ -10,7 +10,7 @@ import AdminPanel from '../admin';
 import NotFound from '../errors/NotFound';
 import Navbar from '../../containers/navbar';
 import DashBoard from '../dashboard';
-import FrontEndComponent from '../frontEndComponents';
+import UIDemo from '../demo';
 
 import './App.sass';
 
@@ -32,7 +32,7 @@ const App = () => (
             <Route path="/logout" component={Logout} />
             <Route path="/dashboard" component={DashBoard} />
             <Route path="/admin" component={AdminPanel} />
-            <Route path="/front_end_components" component={FrontEndComponent} />
+            <Route path="/ui_demo" component={UIDemo} />
             <Route path="/page_not_found" component={NotFound} />
             <Route path="*" component={() => <Redirect to="/page_not_found" />} />
           </Switch>
