+++ web/components/input/buttons/index.js
@@ -1,5 +1,7 @@
 import SecondaryButton from './SecondaryButton/index';
+import PrimaryButton from './PrimaryButton';
 
 export {
   SecondaryButton,
+  PrimaryButton,
 };
